Implement Art Object Details Screen
This commit implements the view controller that presents art object details. It also includes an update for `MainCoordinator` to enable navigation to this screen. Additionally, it introduces a helper for converting hex values to UIColor and includes tests for it. Lastly, it adds the `ErrorModalView`, a supplementary view for displaying errors to the user.